Make a few methods public to become compatible with jruby-9.2.10.0
Also jdk 9.

We can't expose implementation methods anymore, and we also can't send
on private methods. To work around this we'll expose certain methods and
accessors via the interfaces. This is a big hack, but there's already
efforts to upgrade to the new version of nats.java, so this is a "stop
the bleeding" fix.